refactor: introduce ambient_project_macro_common
