+++ shared_crates/project/src/lib.rs
@@ -3,11 +3,6 @@ mod native;
 #[cfg(feature = "native")]
 pub use native::*;
 
-#[cfg(feature = "guest")]
-mod guest;
-#[cfg(feature = "guest")]
-pub use guest::*;
-
 mod component;
 pub use component::*;
 mod concept;
