+++ shared_crates/project_macro_common/rustfmt.toml
@@ -1,0 +1,2 @@
+max_width = 100
+use_small_heuristics = "Default"
