+++ shared_crates/project_macro_common/src/api_project/component.rs
