+++ shared_crates/project_macro_common/src/lib.rs
@@ -1,0 +1,6 @@
+extern crate proc_macro;
+
+mod api_project;
+pub use api_project::*;
+
+pub const MANIFEST: &str = include_str!("../ambient.toml");
